refactor: MainPage 관련 클래스 제거
SearchRequest의 클래스 위치 수정
