+++ backend/server/src/main/java/com/airbnb/controller/RoomController.java
@@ -1,7 +1,7 @@
 package com.airbnb.controller;
 
-import com.airbnb.dto.mainpage.search.SearchRequest;
 import com.airbnb.dto.room.RoomListResponse;
+import com.airbnb.dto.room.SearchRequest;
 import com.airbnb.service.RoomService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
