--- backend/server/src/main/java/com/airbnb/domain/mainpage/search/Calendar.java
@@ -1,20 +1,0 @@
-package com.airbnb.domain.mainpage.search;
-
-public class Calendar {
-
-    private final String[] title;
-    private final String placeholderText;
-
-    public Calendar(String[] title, String placeholderText) {
-        this.title = title;
-        this.placeholderText = placeholderText;
-    }
-
-    public String[] getTitle() {
-        return title;
-    }
-
-    public String getPlaceholderText() {
-        return placeholderText;
-    }
-}
