--- backend/server/src/main/java/com/airbnb/domain/mainpage/search/Guests.java
@@ -1,19 +1,0 @@
-package com.airbnb.domain.mainpage.search;
-
-public class Guests {
-    private final String title;
-    private final String placeholderText;
-
-    public Guests(String title, String placeholderText) {
-        this.title = title;
-        this.placeholderText = placeholderText;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getPlaceholderText() {
-        return placeholderText;
-    }
-}
