--- backend/server/src/main/java/com/airbnb/domain/mainpage/search/SearchInputComponents.java
@@ -1,14 +1,0 @@
-package com.airbnb.domain.mainpage.search;
-
-public class SearchInputComponents {
-
-    private final Location location;
-
-    public SearchInputComponents(Location location) {
-        this.location = location;
-    }
-
-    public Location getLocation() {
-        return location;
-    }
-}
