+++ backend/server/src/main/java/com/airbnb/dto/room/SearchRequest.java
@@ -1,4 +1,4 @@
-package com.airbnb.dto.mainpage.search;
+package com.airbnb.dto.room;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
